pyproject.toml
[project]
name = "sovereign-os"
version = "0.1.0"
description = "Sovereign Agentic OS with HLF"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "redis>=5.2.0",
    "dspy-ai>=2.5.0",
    "lark>=1.2.0",
    "tiktoken>=0.8.0",
    "ulid-py>=1.1.0",
    "httpx>=0.27.0",
    "traceloop-sdk>=0.30.0",
    "alembic>=1.13.0",
    "sqlalchemy>=2.0.0",
    "typer>=0.12.0",
    "docker>=7.1.0",
    "sqlite-vec>=0.1.6",
    "pyarrow>=23.0.1",
    "pystray>=0.19.5",
    "pillow>=12.1.1",
    "streamlit>=1.54.0",
    "plotly>=6.5.2",
    "pandas>=2.3.3",
    "numpy>=2.1.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "gspread>=6.1.0",
    "mcp[cli]>=1.0.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "pytest-asyncio>=0.24.0", "pre-commit>=4.0.0", "ruff>=0.7.0"]

[project.scripts]
hlfc = "hlf.hlfc:main"
hlffmt = "hlf.hlffmt:main"
hlflint = "hlf.hlflint:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 120
extend-exclude = ["hlf/_parser_cache.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "hlf"]
